+++ bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/schuetzenstatistikletztejahre/api/types/SchuetzenstatistikLetzteJahreDO.java
@@ -12,12 +12,12 @@ import de.bogenliga.application.common.component.types.DataObject;
 public class SchuetzenstatistikLetzteJahreDO extends CommonDataObject implements DataObject {
 
     private String schuetzenname;
-    private long sportjahr1;
-    private long sportjahr2;
-    private long sportjahr3;
-    private long sportjahr4;
-    private long sportjahr5;
-    private long allejahre_schnitt;
+    private float sportjahr1;
+    private float sportjahr2;
+    private float sportjahr3;
+    private float sportjahr4;
+    private float sportjahr5;
+    private float allejahre_schnitt;
 
     public SchuetzenstatistikLetzteJahreDO(){
         //default constructor
@@ -36,12 +36,12 @@ public class SchuetzenstatistikLetzteJahreDO extends CommonDataObject implements
      */
     public SchuetzenstatistikLetzteJahreDO(
             String schuetzenname,
-            long sportjahr1,
-            long sportjahr2,
-            long sportjahr3,
-            long sportjahr4,
-            long sportjahr5,
-            long allejahre_schnitt
+            float sportjahr1,
+            float sportjahr2,
+            float sportjahr3,
+            float sportjahr4,
+            float sportjahr5,
+            float allejahre_schnitt
     ) {
         this.schuetzenname = schuetzenname;
         this.sportjahr1 = sportjahr1;
@@ -62,57 +62,57 @@ public class SchuetzenstatistikLetzteJahreDO extends CommonDataObject implements
         this.schuetzenname = schuetzenname;
     }
 
-    public long getSportjahr4() {
+    public float getSportjahr4() {
         return sportjahr4;
     }
 
 
-    public void setSportjahr4(long sportjahr4) {
+    public void setSportjahr4(float sportjahr4) {
         this.sportjahr4 = sportjahr4;
     }
 
 
-    public long getSportjahr1() {
+    public float getSportjahr1() {
         return sportjahr1;
     }
 
 
-    public void setSportjahr1(long sportjahr1) {
+    public void setSportjahr1(float sportjahr1) {
         this.sportjahr1 = sportjahr1;
     }
 
 
-    public long getSportjahr2() {
+    public float getSportjahr2() {
         return sportjahr2;
     }
 
-    public void setAllejahre_schnitt(long allejahre_schnitt) {
+    public void setAllejahre_schnitt(float allejahre_schnitt) {
         this.allejahre_schnitt = allejahre_schnitt;
     }
 
-    public long getAllejahre_schnitt() {
+    public float getAllejahre_schnitt() {
         return allejahre_schnitt;
     }
 
-    public long getSportjahr5() {
+    public float getSportjahr5() {
         return sportjahr5;
     }
 
-    public void setSportjahr5(long sportjahr5) {
+    public void setSportjahr5(float sportjahr5) {
         this.sportjahr5 = sportjahr5;
     }
 
-    public void setSportjahr2(long sportjahr2) {
+    public void setSportjahr2(float sportjahr2) {
         this.sportjahr2 = sportjahr2;
     }
 
 
-    public long getSportjahr3() {
+    public float getSportjahr3() {
         return sportjahr3;
     }
 
 
-    public void setSportjahr3(long sportjahr3) {
+    public void setSportjahr3(float sportjahr3) {
         this.sportjahr3 = sportjahr3;
     }
