+++ bogenliga/bogenliga-business-logic/src/main/java/de/bogenliga/application/business/schuetzenstatistikletztejahre/impl/entity/SchuetzenstatistikLetzteJahreBE.java
@@ -11,12 +11,12 @@ public class SchuetzenstatistikLetzteJahreBE extends CommonBusinessEntity implem
     private static final long serialVersionUID = 5734330182051890905L;
 
     private String schuetzenname;
-    private long sportjahr1;
-    private long sportjahr2;
-    private long sportjahr3;
-    private long sportjahr4;
-    private long sportjahr5;
-    private long allejahre_schnitt;
+    private float sportjahr1;
+    private float sportjahr2;
+    private float sportjahr3;
+    private float sportjahr4;
+    private float sportjahr5;
+    private float allejahre_schnitt;
 
     @Override
     public String toString() {
@@ -41,62 +41,62 @@ public class SchuetzenstatistikLetzteJahreBE extends CommonBusinessEntity implem
     }
 
 
-    public long getAllejahre_schnitt() {
+    public float getAllejahre_schnitt() {
         return allejahre_schnitt;
     }
 
 
-    public void setAllejahre_schnitt(long allejahre_schnitt) {
+    public void setAllejahre_schnitt(float allejahre_schnitt) {
         this.allejahre_schnitt = allejahre_schnitt;
     }
 
 
-    public long getSportjahr1() {
+    public float getSportjahr1() {
         return sportjahr1;
     }
 
 
-    public void setSportjahr1(long sportjahr1) {
+    public void setSportjahr1(float sportjahr1) {
         this.sportjahr1 = sportjahr1;
     }
 
 
-    public long getSportjahr2() {
+    public float getSportjahr2() {
         return sportjahr2;
     }
 
 
-    public void setSportjahr2(long sportjahr2) {
+    public void setSportjahr2(float sportjahr2) {
         this.sportjahr2 = sportjahr2;
     }
 
 
-    public long getSportjahr4() {
+    public float getSportjahr4() {
         return sportjahr4;
     }
 
 
-    public void setSportjahr4(long sportjahr4) {
+    public void setSportjahr4(float sportjahr4) {
         this.sportjahr4 = sportjahr4;
     }
 
 
-    public long getSportjahr3() {
+    public float getSportjahr3() {
         return sportjahr3;
     }
 
 
-    public void setSportjahr3(long sportjahr3) {
+    public void setSportjahr3(float sportjahr3) {
         this.sportjahr3 = sportjahr3;
     }
 
 
-    public long getSportjahr5() {
+    public float getSportjahr5() {
         return sportjahr5;
     }
 
 
-    public void setSportjahr5(long sportjahr5) {
+    public void setSportjahr5(float sportjahr5) {
         this.sportjahr5 = sportjahr5;
     }
